+++ src/backend/templates/InvitationLink.spec.ts
@@ -8,6 +8,7 @@ describe("InvitationLink", () => {
         const result = InvitationLink(Languages.ENGLISH, {
           role: Roles.STUDENT,
           link: "https://mylink.de",
+          username: "ralle",
         });
         expect(result.html).toContain("your missed lessons");
         expect(result.html).not.toContain("to administer");
@@ -20,6 +21,7 @@ describe("InvitationLink", () => {
         const result = InvitationLink(Languages.ENGLISH, {
           role: Roles.MANAGER,
           link: "https://mylink.de",
+          username: "ralle",
         });
         expect(result.html).toContain("to manage your students");
         expect(result.html).not.toContain("your missed lessons");
@@ -33,6 +35,7 @@ describe("InvitationLink", () => {
         const result = InvitationLink(Languages.GERMAN, {
           role: Roles.STUDENT,
           link: "https://mylink.de",
+          username: "ralle",
         });
         expect(result.html).toContain("ihre Fehlstunden");
       });
@@ -43,6 +46,7 @@ describe("InvitationLink", () => {
         const result = InvitationLink(Languages.GERMAN, {
           role: Roles.MANAGER,
           link: "https://mylink.de",
+          username: "ralle",
         });
         expect(result.html).toContain("ihrer Stufe zu verwalten");
       });